+++ scripts/module.js
@@ -24,7 +24,7 @@ Hooks.once("ready", async function () {
 
     const isBindHeroicSpiritSuccess = isBindHeroicSpiritHit(chatMessage);
 
-    if (chatMessage.isRoll) return;
+    if (chatMessage.isRoll && !isBindHeroicSpiritSuccess) return;
 
     // Handle Specific Case Bind Heroic Spirit
     const itemUuid = isBindHeroicSpiritSuccess